+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.3.0 (2016-02-27)
+
+* Added support for `Encoder`/`Decoder`. This provides `ToInput` and `FromInput` instances for arbitrary tuples, case classes, etc. as long
+  as you have appropriate `io.circe.Encoder`/`io.circe.Decoder` in scope (in most cases it's enough to just `import io.circe.generic.auto._`)
+
 ## v0.2.0 (2016-02-15)
 
 * Updated `circe` to version 0.3.0